emu: polish config struct
